Adapt based on artifact availability
When rpc_release is changed, all PR's will fail
until the artifacts for that rpc_release are
generated (which takes ~8 hours).

This will stall development and cause a
chicken-and-egg problem where the PR to change
the release can't merge because the artifacts
the build will try to use do not exist, nor will
they until the patch to change the release
merges.

This patch solves that problem by making the AIO
configuration and deployment process use
artifacts if they're available, or not use them
if they are not available.

The artifact build scripts are adapted to re-use
the same functions to clean up a bit of code
duplication.

(cherry picked from commit 698390d69b9398d631275ee57aec6cdecbe4b00e)