Deploy ironic if the DEPLOY_IRONIC variable is set
We deploy ironic if told to.  If not, do not deploy.  We do this by
assigning the ironic containers to the os-infra-containers groups.  By
default they are assigned to their own group.  We set up the env.d file
in both deploy.sh and boostrap-ironic.yml (called by boostrap-aio.sh)
as just because one may be used does not mean the other one is.