Ensure that the apt artifact repo is configured
When using the apt artifacts, we currently remove
the updates source from /etc/apt/sources.list but
do not ensure that the rpco.list file is configured
prior to executing the openstack_hosts role. This
causes the build to fail because that role needs
packages not available in the base source.

This patch implements the execution of the
pip_install role which happens to also configure
the additional repo needed for everything to be
right with the world again.

(cherry picked from commit cd3543ddaf741edb29b17202242de76aada6913e)